Distance Gap Issue Complete
Touched up sizing for small and mobile devices and imported updated Navbar.
Known issue with footer overlapping page.
